fix bdist_conda to ignore variants - split builds don't work with variants.

+++ conda_build/build.py
@@ -1011,9 +1011,10 @@ def build(m, index, post=None, need_source_download=True, need_reparse_in_env=Fa
                 assert m.final, "output metadata for {} is not finalized".format(m.dist())
                 pkg_path = bldpkg_path(m)
                 if pkg_path not in built_packages and pkg_path not in new_pkgs:
-                    utils.rm_rf(m.config.host_prefix)
-                    utils.rm_rf(m.config.build_prefix)
-                    utils.rm_rf(m.config.test_prefix)
+                    if post is None:
+                        utils.rm_rf(m.config.host_prefix)
+                        utils.rm_rf(m.config.build_prefix)
+                        utils.rm_rf(m.config.test_prefix)
                     sub_host_ms_deps = m.ms_depends('host')
                     if host_index:
                         host_actions = environ.get_install_actions(m.config.host_prefix, host_index,
@@ -1424,10 +1425,15 @@ def build_tree(recipe_list, config, build_only=False, post=False, notest=False,
                     config.index = None
                 metadata_tuples, index = render_recipe(recipe, config=config, variants=variants,
                                                        permit_unsatisfiable_variants=False)
+            # restrict to building only one variant for bdist_conda.  The way it splits the build
+            #    job breaks variants horribly.
+            if post in (True, False):
+                metadata_tuples = metadata_tuples[:1]
             for (metadata, need_source_download, need_reparse_in_env) in metadata_tuples:
-                utils.rm_rf(metadata.config.host_prefix)
-                utils.rm_rf(metadata.config.build_prefix)
-                utils.rm_rf(metadata.config.test_prefix)
+                if post is None:
+                    utils.rm_rf(metadata.config.host_prefix)
+                    utils.rm_rf(metadata.config.build_prefix)
+                    utils.rm_rf(metadata.config.test_prefix)
 
                 if metadata.name() not in metadata.config.build_folder:
                     metadata.config.compute_build_id(metadata.name(), reset=True)
